# apps/api/docs/swagger.yaml
basePath: /
definitions:
...
  events.Event:
    properties:
      content:
        example: 'Issue body: Please update the connector README with setup steps.'
        type: string
      id:
        example: evt_a1b2c3d4e5f6
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      occurred_at:
        example: "2026-05-29T10:00:00Z"
        type: string
      schema_version:
        example: v1
        type: string
      source:
        example: github
        type: string
      source_id:
        example: https://github.com/owner/repo/issues/1
        type: string
      subject:
        example: 'Fix: update connector README'
        type: string
      trace_id:
        example: trace_a1b2c3d4e5f6
        type: string
      type:
        allOf:
        - $ref: '#/definitions/events.Type'
        example: document.ingested
    type: object
  events.Type:
    enum:
    - document.ingested
    - document.normalized
    - entity.extracted
    - identity.resolved
    - relationship.created
    - mismatch.detected
    - codex.analysis.completed
    type: string
    x-enum-varnames:
    - DocumentIngested
    - DocumentNormalized
    - EntityExtracted
    - IdentityResolved
    - RelationshipCreated
    - MismatchDetected
    - CodexAnalysisComplete
...
  request.FilesystemIngest:
    properties:
      content:
        example: Optional raw content instead of reading from URI
        type: string
      cursor:
        example: eyJsYXN0X2lkIjoiMTIzIn0=
        type: string
      exclude:
        example: node_modules,*.log
        type: string
      include:
        example: '*.go,*.md'
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      uri:
        example: storage/raw/README.md
        type: string
      workspace_id:
        example: /home/user/myproject
        type: string
    type: object
...
  request.NotionIngest:
    properties:
      content:
        example: Optional raw content instead of fetching from Notion
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      provider:
        example: token
        type: string
      token:
        example: secret_xxxx
        type: string
      uri:
        example: notion://page/PAGE_ID
        type: string
      workspace_id:
        example: /home/user/myproject
        type: string
    type: object
...
  request.SharePointIngest:
    properties:
      client_id:
        example: 11111111-1111-1111-1111-111111111111
        type: string
      client_secret:
        example: secret~value
        type: string
      content:
        example: Optional raw content instead of fetching from Graph
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      provider:
        example: token
        type: string
      tenant_id:
        example: 00000000-0000-0000-0000-000000000000
        type: string
      token:
        example: eyJ0...
        type: string
      uri:
        example: sharepoint://sites/mysite/items/abcdef01
        type: string
      workspace_id:
        example: /home/user/myproject
        type: string
    type: object
...
  response.Ingest:
    properties:
      capabilities:
        items:
          type: string
        type: array
      connector:
        example: github
        type: string
      entity_count:
        type: integer
      event:
        $ref: '#/definitions/events.Event'
      event_count:
        example: 1
        type: integer
      events:
        items:
          $ref: '#/definitions/events.Event'
        type: array
      metadata:
        additionalProperties:
          type: string
        type: object
      metadata_items:
        items:
          additionalProperties:
            type: string
          type: object
        type: array
      mismatch_count:
        type: integer
      persisted_event_count:
        type: integer
      persistence_mode:
        type: string
      preview:
        example: 'Issue #1: Fix connector README — requesting updated setup steps.'
        type: string
      previews:
        items:
          type: string
        type: array
      relationship_count:
        type: integer
      workspace_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Local-first pipeline API for ingesting and reasoning over engineering
    context.
  title: ContextOS API
  version: "1.0"
paths:
...
  /filesystem/ingest:
    post:
      consumes:
      - application/json
      description: Reads a local file or folder path and returns provenance-rich events
        with extracted text and file metadata.
      parameters:
      - description: Filesystem ingest request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.FilesystemIngest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Ingest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ingest a local filesystem artifact
      tags:
      - filesystem
...
  /notion/ingest:
    post:
      consumes:
      - application/json
      description: Fetches Notion page blocks or database entries by URI and returns
        a provenance-rich event. Set provider=codex to route through the Notion Codex
        plugin.
      parameters:
      - description: Notion ingest request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.NotionIngest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Ingest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ingest a Notion page or database
      tags:
      - notion
  /notion/ingest/stream:
    post:
      consumes:
      - application/json
      description: Streams Notion Codex plugin progress via SSE, then emits a result
        event.
      parameters:
      - description: Notion ingest request (provider must be codex)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.NotionIngest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stream Notion Codex ingest
      tags:
      - notion
  /notion/status:
    get:
      description: Returns whether a Notion integration token environment variable
        is configured.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Notion connection status
      tags:
      - notion
...
  /sharepoint/ingest:
    post:
      consumes:
      - application/json
      description: Fetches a SharePoint or OneDrive file by URI using Microsoft Graph
        and returns a provenance-rich event. Set provider=codex to route through the
        SharePoint Codex plugin.
      parameters:
      - description: SharePoint ingest request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.SharePointIngest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Ingest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Ingest a SharePoint / OneDrive item
      tags:
      - sharepoint
  /sharepoint/ingest/stream:
    post:
      consumes:
      - application/json
      description: Streams SharePoint Codex plugin progress via SSE, then emits a
        result event.
      parameters:
      - description: SharePoint ingest request (provider must be codex)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.SharePointIngest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stream SharePoint Codex ingest
      tags:
      - sharepoint
  /sharepoint/status:
    get:
      description: Returns whether a SharePoint access token or client credentials
        are configured for Microsoft Graph ingest.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: SharePoint connection status
      tags:
      - sharepoint
...
swagger: "2.0"